fix(streaming/five): fileIdx int16 clamping for mod raw streamer

+++ code/components/gta-streaming-five/src/Streaming.cpp
@@ -213,7 +213,7 @@ namespace streaming
 		if (GetRawStreamerForFile(fileName, &rawStreamer))
 		{
 			auto fileIdx = rawStreamer->GetEntryByName(fileName);
-			if (fileIdx != -1)
+			if (fileIdx != uint16_t(-1))
 			{
 				const char* registerName = fileName;
